WoT Group C (C1): extract getCellCoordsFromMousePosition from event.ts into utils/hitTest.ts
