+++ handsontable/src/3rdparty/walkontable/src/event.ts
@@ -9,8 +9,7 @@ import {
   getParent,
 } from '../../../helpers/dom/element';
 import { partial } from '../../../helpers/function';
-import { clamp } from '../../../helpers/number';
-import { findColumnAtX, findRowAtY } from './utils/hitTest';
+import { getCellCoordsFromMousePosition } from './utils/hitTest';
 import { isTouchSupported } from '../../../helpers/feature';
 import { isMobileBrowser, isChromeWebKit, isFirefoxWebKit, isIOS } from '../../../helpers/browser';
 import { isDefined } from '../../../helpers/mixed';
@@ -420,161 +419,7 @@ class Event {
    * @returns {{ coords: CellCoords, isOutside: boolean }}
    */
   #getCellCoordsFromMousePosition(mouseX: number, mouseY: number) {
-    const isRtl = this.#deps.wtSettings.getSetting<boolean>('rtlMode');
-    // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-    const wot = this.#deps.facadeGetter();
-
-    const numberOfFixedColumnsStart = this.#deps.wtSettings.getSetting<number>('fixedColumnsStart');
-    const numberOfFixedRowsTop = this.#deps.wtSettings.getSetting<number>('fixedRowsTop');
-    const numberOfFixedRowsBottom = this.#deps.wtSettings.getSetting<number>('fixedRowsBottom');
-
-    // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-    const firstPartiallyVisibleRow: number = wot.wtScroll.getFirstPartiallyVisibleRow();
-    // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-    const lastPartiallyVisibleRow: number = wot.wtScroll.getLastPartiallyVisibleRow();
-    // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-    const firstPartiallyVisibleColumn: number = wot.wtScroll.getFirstPartiallyVisibleColumn();
-    // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-    const lastPartiallyVisibleColumn: number = wot.wtScroll.getLastPartiallyVisibleColumn();
-    const tableOffset = this.#deps.geometryReader.getBoundingClientRect(this.#deps.wtTable.wtRootElement);
-
-    // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-    const columnHeaderHeight: number = this.#deps.wtSettings.getSetting<Function[]>('columnHeaders').length > 0
-      ? wot.wtViewport.getColumnHeaderHeight() : 0;
-    // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-    const rowHeaderWidth: number = this.#deps.wtSettings.getSetting<Function[]>('rowHeaders').length > 0
-      ? wot.wtViewport.getRowHeaderWidth() : 0;
-    const rootWindow = this.#deps.rootWindow;
-    // When the window is the scroll container and tableOffset.left/top > 0 (e.g. RTL
-    // at max-left scroll where tableOffset.left can exceed innerWidth), using it as the
-    // clamp minimum causes clamp(min > max) to always return min, mapping every mouse
-    // position to the wrong edge column. Math.min(0, tableOffset) corrects this while
-    // preserving the original boundary when the table is partially off-screen to the
-    // left/top (tableOffset < 0), which is the normal scrolled-past-origin case.
-    const tableViewportLeft = wot.wtViewport.isHorizontallyScrollableByWindow()
-      ? Math.min(0, tableOffset.left)
-      : tableOffset.left;
-    // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-    const tableViewportTop: number = wot.wtViewport.isVerticallyScrollableByWindow()
-      ? Math.min(0, tableOffset.top)
-      : tableOffset.top;
-    // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-    const tableViewportRight: number = wot.wtViewport.isHorizontallyScrollableByWindow()
-      ? rootWindow.innerWidth
-      : tableOffset.left + wot.wtViewport.getViewportWidth() + rowHeaderWidth;
-    // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-    const tableViewportBottom: number = wot.wtViewport.isVerticallyScrollableByWindow()
-      ? rootWindow.innerHeight
-      : tableOffset.top + wot.wtViewport.getViewportHeight() + columnHeaderHeight;
-
-    const clampedX = clamp(mouseX, tableViewportLeft, tableViewportRight);
-    const clampedY = clamp(mouseY, tableViewportTop, tableViewportBottom);
-
-    let foundColumn = null;
-
-    if (numberOfFixedColumnsStart > 0) {
-      // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-      const fixedCell = wot.getCell({ row: firstPartiallyVisibleRow, col: 0 }, true);
-
-      if (fixedCell instanceof HTMLElement) {
-        const fixedCellRect = this.#deps.geometryReader.getBoundingClientRect(fixedCell);
-        const fixedRelativeX = isRtl ? fixedCellRect.right - clampedX : clampedX - fixedCellRect.left;
-
-        foundColumn = findColumnAtX(wot, firstPartiallyVisibleRow, 0, numberOfFixedColumnsStart - 1, fixedRelativeX);
-      }
-    }
-
-    if (foundColumn === null) {
-      // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-      const scrollCell = wot.getCell({ row: firstPartiallyVisibleRow, col: firstPartiallyVisibleColumn }, true);
-
-      if (scrollCell instanceof HTMLElement) {
-        const scrollCellRect = this.#deps.geometryReader.getBoundingClientRect(scrollCell);
-        const scrollRelativeX = isRtl ? scrollCellRect.right - clampedX : clampedX - scrollCellRect.left;
-
-        foundColumn = findColumnAtX(
-          wot,
-          firstPartiallyVisibleRow,
-          firstPartiallyVisibleColumn,
-          lastPartiallyVisibleColumn,
-          scrollRelativeX,
-        );
-
-        if (foundColumn === null) {
-          foundColumn = scrollRelativeX < 0 ? firstPartiallyVisibleColumn : lastPartiallyVisibleColumn;
-        }
-      } else {
-        foundColumn = firstPartiallyVisibleColumn;
-      }
-    }
-
-    let foundRow = null;
-
-    if (numberOfFixedRowsTop > 0) {
-      // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-      const fixedCell = wot.getCell({ row: 0, col: firstPartiallyVisibleColumn }, true);
-
-      if (fixedCell instanceof HTMLElement) {
-        const fixedCellRect = this.#deps.geometryReader.getBoundingClientRect(fixedCell);
-        const fixedRelativeY = clampedY - fixedCellRect.top;
-
-        foundRow = findRowAtY(wot, firstPartiallyVisibleColumn, 0, numberOfFixedRowsTop - 1, fixedRelativeY);
-      }
-    }
-
-    if (foundRow === null && numberOfFixedRowsBottom > 0) {
-      const totalRows = this.#deps.wtSettings.getSetting<number>('totalRows');
-      const bottomStartRow = totalRows - numberOfFixedRowsBottom;
-      const bottomEndRow = totalRows - 1;
-      // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-      const fixedBottomCell = wot.getCell({ row: bottomStartRow, col: firstPartiallyVisibleColumn }, true);
-
-      if (fixedBottomCell instanceof HTMLElement) {
-        const fixedBottomCellRect = this.#deps.geometryReader.getBoundingClientRect(fixedBottomCell);
-        const fixedBottomRelativeY = clampedY - fixedBottomCellRect.top;
-
-        if (fixedBottomRelativeY >= 0) {
-          foundRow = findRowAtY(wot, firstPartiallyVisibleColumn, bottomStartRow, bottomEndRow, fixedBottomRelativeY);
-
-          if (foundRow === null) {
-            foundRow = bottomEndRow;
-          }
-        }
-      }
-    }
-
-    if (foundRow === null) {
-      // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-      const scrollCell = wot.getCell({ row: firstPartiallyVisibleRow, col: firstPartiallyVisibleColumn }, true);
-
-      if (scrollCell instanceof HTMLElement) {
-        const scrollCellRect = this.#deps.geometryReader.getBoundingClientRect(scrollCell);
-        const scrollRelativeY = clampedY - scrollCellRect.top;
-
-        foundRow = findRowAtY(
-          wot,
-          firstPartiallyVisibleColumn,
-          firstPartiallyVisibleRow,
-          lastPartiallyVisibleRow,
-          scrollRelativeY,
-        );
-
-        if (foundRow === null) {
-          foundRow = lastPartiallyVisibleRow;
-        }
-      } else {
-        foundRow = firstPartiallyVisibleRow;
-      }
-    }
-
-    return {
-      // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-      coords: wot.createCellCoords(foundRow, foundColumn),
-      isOutside: mouseX < tableViewportLeft ||
-                 mouseX > tableViewportRight ||
-                 mouseY < tableViewportTop ||
-                 mouseY > tableViewportBottom,
-    };
+    return getCellCoordsFromMousePosition(this.#deps, mouseX, mouseY);
   }
 
   /**
